fix(review-eval): the deadline reaches a stalled capped stream; split out the schedule-issue mode
`run_stream_capped` starts the model under job control, which puts it in a
process group the `run_bounded` watchdog does not signal. A codex that stalled
before writing the bytes `head` waits for therefore outlived the deadline: the
TERM killed the reader, the run reported failure, and the model kept spending
against its own hour-long timeout. TERM and INT handlers now KILL the child's
group and exit 143, and every return path clears them. A new shell-lifecycle
test runs the committed functions and proves a silent, sleeping child is gone
within seconds of the deadline; without the handlers it keeps its heartbeat.

`review-eval.mjs` stood at 899 lines against a 900-line pin, and the recurring
review checklist says a file already past the 600-line soft cap must be split
rather than grown. The `--schedule-issue` mode moves to
`review-eval-schedule-issue.mjs`, which imports from the defining modules
instead of `review-eval.mjs` or the run barrel so no cycle forms.
`review-eval.mjs` re-exports `runScheduleIssue`, so
`review-eval-freshness-publication.mjs` is unchanged. The file is now 809
lines and its pin drops to 850.
